melengkapi fitur admin (belum revisi)

+++ routes/web.php
@@ -14,11 +14,10 @@
 // Route::get('/', function () {
 //     return view('welcome');
 // });
-Route::get('index/login', function(){
-	return view('admin.login');
-});
+Route::get('index/login', 'Admin\LoginController@index');
+Route::post('index/admin', 'Admin\LoginController@login');
 
-Route::prefix('index/admin')->group(function(){
+Route::middleware(['auth:admin'])->prefix('index/admin')->group(function(){
 
 	Route::get('dashboard', function(){
 		return view('admin.dashboard');
@@ -28,45 +27,53 @@ Route::prefix('index/admin')->group(function(){
 	Route::get('ajax/get_kota/{province_id}', 'Admin\AjaxController@get_kota');
 	Route::get('ajax/get_kecamatan/{regency_id}', 'Admin\AjaxController@get_kecamatan');
 	Route::get('ajax/get_kelurahan/{district_id}', 'Admin\AjaxController@get_kelurahan');
+	Route::get('ajax/get_produk/{tipe_produk}', 'Admin\AjaxController@get_produk');
+	Route::get('ajax/get_user/{tipe_produk}', 'Admin\AjaxController@get_user');
 	Route::get('ajax/detail_karyawan/{id_karyawan}', 'Admin\AjaxController@detail_karyawan');
+	Route::get('ajax/get_transaksi/{keyword}', 'Admin\AjaxController@get_transaksi');
 
 	Route::get('transaksi/haji', 'Admin\TransaksiController@haji');
-	Route::post('transaksi/haji/konfirmasi/{id}', 'Admin\TransaksiController@konfirm');
+	Route::post('transaksi/haji/konfirmasi/{id}', 'Admin\TransaksiController@konfirm')->middleware(['permission:konfirmasi transaksi']);
 	Route::get('transaksi/umroh', 'Admin\TransaksiController@umroh');
-	Route::post('transaksi/umroh/konfirmasi/{id}', 'Admin\TransaksiController@konfirm');
+	Route::post('transaksi/umroh/konfirmasi/{id}', 'Admin\TransaksiController@konfirm')->middleware(['permission:konfirmasi transaksi']);
 	Route::get('transaksi/wisata', 'Admin\TransaksiController@wisata');
-	Route::post('transaksi/wisata/konfirmasi/{id}', 'Admin\TransaksiController@konfirm');
+	Route::post('transaksi/wisata/konfirmasi/{id}', 'Admin\TransaksiController@konfirm')->middleware(['permission:konfirmasi transaksi']);
 	Route::get('transaksi/sedekah', 'Admin\TransaksiController@sedekah');
-	Route::post('transaksi/sedekah/konfirmasi/{id}', 'Admin\TransaksiController@konfirm');
+	Route::post('transaksi/sedekah/konfirmasi/{id}', 'Admin\TransaksiController@konfirm')->middleware(['permission:konfirmasi transaksi']);
 	Route::get('transaksi/bayar-paket', 'Admin\TransaksiController@bayar_paket');
-	Route::post('transaksi/bayar-paket/konfirmasi/{id}', 'Admin\TransaksiController@konfirm_paket');
+	Route::post('transaksi/bayar-paket/konfirmasi/{id}', 'Admin\TransaksiController@konfirm_paket')->middleware(['permission:konfirmasi transaksi']);
 
-	Route::get('produk/tambah', 'Admin\ProdukController@create');
+	Route::get('kwitansi', 'Admin\KwitansiController@index');
+	Route::get('kwitansi/buat-transaksi', 'Admin\KwitansiController@buat_transaksi')->middleware(['permission:buat transaksi']);
+	Route::post('kwitansi/buat-transaksi/simpan', 'Admin\KwitansiController@store_transaksi');
+	Route::post('kwitansi/status/{id_kwitansi}', 'Admin\KwitansiController@ubah_status')->middleware(['permission:ubah status kwitansi']);
+
+	Route::get('produk/tambah', 'Admin\ProdukController@create')->middleware(['permission:tambah produk']);
 	Route::post('produk/tambah/simpan', 'Admin\ProdukController@store');
 	Route::get('produk/{produk}', 'Admin\ProdukController@index');
 	Route::get('produk/{produk}/search', 'Admin\ProdukController@search');
-	Route::get('produk/edit/{id_produk}', 'Admin\ProdukController@edit');
+	Route::get('produk/edit/{id_produk}', 'Admin\ProdukController@edit')->middleware(['permission:edit produk']);
 	Route::patch('produk/update/{id_produk}', 'Admin\ProdukController@update');
-	Route::delete('produk/delete/{id_produk}', 'Admin\ProdukController@delete');
+	Route::delete('produk/delete/{id_produk}', 'Admin\ProdukController@delete')->middleware(['permission:delete produk']);
 
 	Route::get('paket', 'Admin\PaketController@index');
-	Route::get('paket/tambah', 'Admin\PaketController@create');
+	Route::get('paket/tambah', 'Admin\PaketController@create')->middleware(['permission:tambah paket haji umroh']);
 	Route::post('paket/tambah/simpan', 'Admin\PaketController@store');
-	Route::get('paket/edit/{id_produk}', 'Admin\PaketController@edit');
+	Route::get('paket/edit/{id_produk}', 'Admin\PaketController@edit')->middleware(['permission:edit paket haji umroh']);;
 	Route::patch('paket/update/{id_produk}', 'Admin\PaketController@update');
-	Route::delete('paket/delete/{id_produk}', 'Admin\PaketController@delete');
-	Route::get('paket/status/{id_produk}', 'Admin\PaketController@status');
+	Route::delete('paket/delete/{id_produk}', 'Admin\PaketController@delete')->middleware(['permission:delete paket haji umroh']);;
+	Route::get('paket/status/{id_produk}', 'Admin\PaketController@status')->middleware(['permission:status paket haji umroh']);;
 
 	Route::get('voucher', 'Admin\VoucherController@index');
 	Route::get('voucher/search', 'Admin\VoucherController@search');
-	Route::get('voucher/tambah', 'Admin\VoucherController@create');
+	Route::get('voucher/tambah', 'Admin\VoucherController@create')->middleware(['permission:tambah voucher']);
 	Route::post('voucher/tambah/simpan', 'Admin\VoucherController@store');
-	Route::get('voucher/edit/{id_voucher}', 'Admin\VoucherController@edit');
+	Route::get('voucher/edit/{id_voucher}', 'Admin\VoucherController@edit')->middleware(['permission:edit voucher']);
 	Route::patch('voucher/update/{id_voucher}', 'Admin\VoucherController@update');
-	Route::delete('voucher/delete/{id_voucher}', 'Admin\VoucherController@delete');
+	Route::delete('voucher/delete/{id_voucher}', 'Admin\VoucherController@delete')->middleware(['permission:delete voucher']);
 
 	Route::get('data-user/agen', 'Admin\DataUserController@index_agen');
-	Route::post('data-user/agen/konfirmasi/{id}', 'Admin\DataUserController@konfirmasi');
+	Route::post('data-user/agen/konfirmasi/{id}', 'Admin\DataUserController@konfirmasi')->middleware(['permission:konfirmasi agen']);
 	Route::get('data-user/agen/search', 'Admin\DataUserController@search_agen');
 	Route::get('data-user/agen/{id_agen}/list-transaksi', 'Admin\DataUserController@list_trx_agen');
 	Route::get('data-user/agen/{id_agen}/list-transaksi/search', 'Admin\DataUserController@search_transaksi');
@@ -77,83 +84,123 @@ Route::prefix('index/admin')->group(function(){
 
 	Route::get('data-booking/jamaah', 'Admin\JamaahController@index');
 	Route::get('data-booking/jamaah/search', 'Admin\JamaahController@search');
-	Route::get('data-booking/jamaah/tambah', 'Admin\JamaahController@create');
+	Route::get('data-booking/jamaah/tambah', 'Admin\JamaahController@create')->middleware(['permission:tambah jamaah']);
 	Route::post('data-booking/jamaah/tambah/simpan', 'Admin\JamaahController@store');
-	Route::get('data-booking/jamaah/edit/{id_jamaah}', 'Admin\JamaahController@edit');
+	Route::get('data-booking/jamaah/edit/{id_jamaah}', 'Admin\JamaahController@edit')->middleware(['permission:edit jamaah']);
 	Route::patch('data-booking/jamaah/update/{id_jamaah}', 'Admin\JamaahController@update');
-	Route::delete('data-booking/jamaah/delete/{id_jamaah}', 'Admin\JamaahController@delete');
+	Route::delete('data-booking/jamaah/delete/{id_jamaah}', 'Admin\JamaahController@delete')->middleware(['permission:delete jamaah']);
 
 	Route::get('data-booking/booking', 'Admin\BookingController@index');
 	Route::get('data-booking/booking/search', 'Admin\BookingController@search');
-	Route::get('data-booking/booking/edit/{id_booking}', 'Admin\BookingController@edit');
+	Route::get('data-booking/booking/edit/{id_booking}', 'Admin\BookingController@edit')->middleware(['permission:edit booking']);
 	Route::patch('data-booking/booking/update/{id_booking}', 'Admin\BookingController@update');
-	Route::get('data-booking/booking/print-voucher/{id_booking}', 'Admin\BookingController@print');
+	Route::get('data-booking/booking/print-voucher/{id_booking}', 'Admin\BookingController@print')->middleware(['permission:print voucher']);
 
-	Route::get('data-booking/paspor/edit/{id_paspor}', 'Admin\PasporController@edit');
+	Route::get('data-booking/paspor/edit/{id_paspor}', 'Admin\PasporController@edit')->middleware(['permission:edit paspor']);
 	Route::patch('data-booking/paspor/update/{id_paspor}', 'Admin\PasporController@update');
 
 	Route::get('data-kloter/kloter', 'Admin\KloterController@index');
 	Route::get('data-kloter/kloter/search', 'Admin\KloterController@search');
-	Route::get('data-kloter/kloter/tambah', 'Admin\KloterController@create');
+	Route::get('data-kloter/kloter/tambah', 'Admin\KloterController@create')->middleware(['permission:tambah kloter']);
 	Route::post('data-kloter/kloter/tambah/simpan', 'Admin\KloterController@store');
-	Route::get('data-kloter/kloter/edit/{id_kloter}', 'Admin\KloterController@edit');
+	Route::get('data-kloter/kloter/edit/{id_kloter}', 'Admin\KloterController@edit')->middleware(['permission:edit kloter']);
 	Route::patch('data-kloter/kloter/update/{id_kloter}', 'Admin\KloterController@update');
-	Route::delete('data-kloter/kloter/delete/{id_kloter}', 'Admin\KloterController@delete');
+	Route::delete('data-kloter/kloter/delete/{id_kloter}', 'Admin\KloterController@delete')->middleware(['permission:delete kloter']);
 
 	Route::get('data-kloter/kloter/list-jamaah/{id_kloter}', 'Admin\KloterController@list_jamaah');
-	Route::get('data-kloter/kloter/list-jamaah/{id_kloter}/tambah', 'Admin\KloterController@isi_kuota');
+	Route::get('data-kloter/kloter/list-jamaah/{id_kloter}/tambah', 'Admin\KloterController@isi_kuota')->middleware(['permission:isi kuota']);
 	Route::post('data-kloter/kloter/list-jamaah/{id_kloter}/tambah/simpan', 'Admin\KloterController@store_isi_kuota');
-	Route::delete('data-kloter/kloter/list-jamaah/{id_kloter}/delete', 'Admin\KloterController@delete_kuota');
+	Route::delete('data-kloter/kloter/list-jamaah/{id_kloter}/delete', 'Admin\KloterController@delete_kuota')->middleware(['permission:delete kuota']);
 
 	Route::get('data-kloter/bus', 'Admin\BusController@index');
 	Route::get('data-kloter/bus/search', 'Admin\BusController@search');
-	Route::get('data-kloter/bus/tambah', 'Admin\BusController@create');
+	Route::get('data-kloter/bus/tambah', 'Admin\BusController@create')->middleware(['permission:tambah bus']);
 	Route::post('data-kloter/bus/tambah/simpan', 'Admin\BusController@store');
-	Route::get('data-kloter/bus/edit/{id_bus}', 'Admin\BusController@edit');
+	Route::get('data-kloter/bus/edit/{id_bus}', 'Admin\BusController@edit')->middleware(['permission:edit bus']);
 	Route::patch('data-kloter/bus/update/{id_bus}', 'Admin\BusController@update');
-	Route::delete('data-kloter/bus/delete/{id_bus}', 'Admin\BusController@delete');
+	Route::delete('data-kloter/bus/delete/{id_bus}', 'Admin\BusController@delete')->middleware(['permission:delete bus']);
 
 	Route::get('data-kloter/bus/list-jamaah/{id_bus}', 'Admin\BusController@list_jamaah');
-	Route::get('data-kloter/bus/list-jamaah/{id_bus}/tambah', 'Admin\BusController@isi_kuota');
+	Route::get('data-kloter/bus/list-jamaah/{id_bus}/tambah', 'Admin\BusController@isi_kuota')->middleware(['permission:isi kuota']);
 	Route::post('data-kloter/bus/list-jamaah/{id_bus}/tambah/simpan', 'Admin\BusController@store_isi_kuota');
-	Route::delete('data-kloter/bus/list-jamaah/{id_bus}/delete', 'Admin\BusController@delete_kuota');
+	Route::delete('data-kloter/bus/list-jamaah/{id_bus}/delete', 'Admin\BusController@delete_kuota')->middleware(['permission:delete kuota']);
 
 	Route::get('data-kloter/kamar', 'Admin\KamarController@index');
 	Route::get('data-kloter/kamar/search', 'Admin\KamarController@search');
-	Route::get('data-kloter/kamar/tambah', 'Admin\KamarController@create');
+	Route::get('data-kloter/kamar/tambah', 'Admin\KamarController@create')->middleware(['permission:tambah kamar']);
 	Route::post('data-kloter/kamar/tambah/simpan', 'Admin\KamarController@store');
-	Route::get('data-kloter/kamar/edit/{id_kamar}', 'Admin\KamarController@edit');
+	Route::get('data-kloter/kamar/edit/{id_kamar}', 'Admin\KamarController@edit')->middleware(['permission:edit kamar']);
 	Route::patch('data-kloter/kamar/update/{id_kamar}', 'Admin\KamarController@update');
-	Route::delete('data-kloter/kamar/delete/{id_kamar}', 'Admin\KamarController@delete');
+	Route::delete('data-kloter/kamar/delete/{id_kamar}', 'Admin\KamarController@delete')->middleware(['permission:delete kamar']);
 
 	Route::get('data-kloter/kamar/list-jamaah/{id_kamar}', 'Admin\KamarController@list_jamaah');
-	Route::get('data-kloter/kamar/list-jamaah/{id_kamar}/tambah', 'Admin\KamarController@isi_kuota');
+	Route::get('data-kloter/kamar/list-jamaah/{id_kamar}/tambah', 'Admin\KamarController@isi_kuota')->middleware(['permission:isi kuota']);
 	Route::post('data-kloter/kamar/list-jamaah/{id_kamar}/tambah/simpan', 'Admin\KamarController@store_isi_kuota');
-	Route::delete('data-kloter/kamar/list-jamaah/{id_kamar}/delete', 'Admin\KamarController@delete_kuota');
+	Route::delete('data-kloter/kamar/list-jamaah/{id_kamar}/delete', 'Admin\KamarController@delete_kuota')->middleware(['permission:delete kuota']);
 
 	Route::get('karyawan', 'Admin\KaryawanController@index');
-	Route::get('karyawan/tambah', 'Admin\KaryawanController@create');
+	Route::get('karyawan/tambah', 'Admin\KaryawanController@create')->middleware(['permission:tambah karyawan']);
 	Route::post('karyawan/tambah/simpan', 'Admin\KaryawanController@store');
-	Route::get('karyawan/edit/{id_karyawan}', 'Admin\KaryawanController@edit');
+	Route::get('karyawan/edit/{id_karyawan}', 'Admin\KaryawanController@edit')->middleware(['permission:edit karyawan']);
 	Route::patch('karyawan/update/{id_karyawan}', 'Admin\KaryawanController@update');
-	Route::delete('karyawan/delete/{id_karyawan}', 'Admin\KaryawanController@delete');
+	Route::delete('karyawan/delete/{id_karyawan}', 'Admin\KaryawanController@delete')->middleware(['permission:delete karyawan']);
 
 	Route::get('divisi', 'Admin\DivisiController@index');
-	Route::get('divisi/tambah', 'Admin\DivisiController@create');
+	Route::get('divisi/tambah', 'Admin\DivisiController@create')->middleware(['permission:tambah divisi']);
 	Route::post('divisi/tambah/simpan', 'Admin\DivisiController@store');
-	Route::get('divisi/edit/{kode_divisi}', 'Admin\DivisiController@edit');
+	Route::get('divisi/edit/{kode_divisi}', 'Admin\DivisiController@edit')->middleware(['permission:edit divisi']);
 	Route::patch('divisi/update/{kode_divisi}', 'Admin\DivisiController@update');
-	Route::delete('divisi/delete/{kode_divisi}', 'Admin\DivisiController@delete');
+	Route::delete('divisi/delete/{kode_divisi}', 'Admin\DivisiController@delete')->middleware(['permission:delete divisi']);
 
 	Route::get('jabatan', 'Admin\JabatanController@index');
-	Route::get('jabatan/tambah', 'Admin\JabatanController@create');
+	Route::get('jabatan/tambah', 'Admin\JabatanController@create')->middleware(['permission:tambah jabatan']);
 	Route::post('jabatan/tambah/simpan', 'Admin\JabatanController@store');
-	Route::get('jabatan/edit/{kode_jabatan}', 'Admin\JabatanController@edit');
+	Route::get('jabatan/edit/{kode_jabatan}', 'Admin\JabatanController@edit')->middleware(['permission:edit jabatan']);
 	Route::patch('jabatan/update/{kode_jabatan}', 'Admin\JabatanController@update');
-	Route::delete('jabatan/delete/{kode_jabatan}', 'Admin\JabatanController@delete');
+	Route::delete('jabatan/delete/{kode_jabatan}', 'Admin\JabatanController@delete')->middleware(['permission:delete jabatan']);
 
-	Route::get('pengaturan/edit-halaman-bantuan', 'Admin\PengaturanController@edit_bantuan');
+	Route::get('report-admin', 'Admin\LogController@index');
+	Route::get('report-admin/rincian-log/{id_admin}', 'Admin\LogController@rincian');
+
+	Route::get('kas', 'Admin\KasController@index');
+	Route::get('kas/search', 'Admin\KasController@search');
+	Route::get('kas/tambah', 'Admin\KasController@create')->middleware(['permission:tambah kas']);
+	Route::post('kas/tambah/simpan', 'Admin\KasController@store');
+	Route::get('kas/edit/{id_kas}', 'Admin\KasController@edit')->middleware(['permission:edit kas']);
+	Route::patch('kas/update/{id_kas}', 'Admin\KasController@update');
+	Route::delete('kas/delete/{id_kas}', 'Admin\KasController@delete')->middleware(['permission:delete kas']);
+
+
+	Route::get('user', 'Admin\AdminController@index');
+	Route::get('user/{id_karyawan}/tambah', 'Admin\AdminController@create')->middleware(['permission:tambah admin']);
+	Route::post('user/{id_karyawan}/tambah/simpan', 'Admin\AdminController@store');
+	Route::get('user/edit/{id_admin}', 'Admin\AdminController@edit')->middleware(['permission:edit admin']);
+	Route::patch('user/update/{id_admin}', 'Admin\AdminController@update');
+	Route::delete('user/delete/{id_admin}', 'Admin\AdminController@delete')->middleware(['permission:delete admin']);
+
+	Route::get('role', 'Admin\RoleController@index');
+	Route::get('role/tambah', 'Admin\RoleController@create')->middleware(['permission:tambah role']);
+	Route::post('role/tambah/simpan', 'Admin\RoleController@store');
+	Route::get('role/edit/{id_role}', 'Admin\RoleController@edit')->middleware(['permission:edit role']);
+	Route::patch('role/update/{id_role}', 'Admin\RoleController@update');
+	Route::delete('role/delete/{id_role}', 'Admin\RoleController@delete')->middleware(['permission:delete role']);
+
+	Route::get('permission', 'Admin\PermissionController@index');
+	Route::get('permission/tambah', 'Admin\PermissionController@create')->middleware(['permission:tambah permission']);
+	Route::post('permission/tambah/simpan', 'Admin\PermissionController@store');
+	Route::delete('permission/delete/{id_permission}', 'Admin\PermissionController@delete')->middleware(['permission:delete permission']);
+
+	Route::get('pengaturan/edit-halaman-bantuan', 'Admin\PengaturanController@edit_bantuan')->middleware(['permission:menu edit bantuan']);
 	Route::patch('pengaturan/edit-halaman-bantuan/update', 'Admin\PengaturanController@update_bantuan');
+
+	Route::get('pengaturan/informasi', 'Admin\InformasiController@index');
+	Route::get('pengaturan/informasi/tambah', 'Admin\InformasiController@create')->middleware(['permission:tambah informasi']);
+	Route::post('pengaturan/informasi/tambah/simpan', 'Admin\InformasiController@store');
+	Route::get('pengaturan/informasi/edit/{id_informasi}', 'Admin\InformasiController@edit')->middleware(['permission:edit informasi']);
+	Route::patch('pengaturan/informasi/update/{id_informasi}', 'Admin\InformasiController@update');
+	Route::delete('pengaturan/informasi/delete/{id_informasi}', 'Admin\InformasiController@delete')->middleware(['permission:delete informasi']);
+
+	Route::post('logout', 'Admin\LoginController@logout');
 });
 
 Route::group(['middleware'=>'guest'], function() {
